FCN update for PIG

+++ functions.py
@@ -313,8 +313,8 @@ class CNN_PIG_Dataset(Dataset):
                 [input0, output0] = pickle.load(file)
             
             rate = int(filename.split("_r")[1][:3])
-            input0 = torch.tensor(input0)
-            output0 = torch.tensor(output0)
+            input0 = torch.tensor(input0, dtype=torch.float32)
+            output0 = torch.tensor(output0, dtype=torch.float32)
             
             if first:
                 self.input = input0
